+++ feincms/contrib/fields.py
@@ -2,13 +2,13 @@ from __future__ import absolute_import, unicode_literals
 
 import json
 import logging
+import six
 from distutils.version import LooseVersion
 
 from django import get_version
 from django import forms
 from django.db import models
 from django.core.serializers.json import DjangoJSONEncoder
-from django.utils import six
 
 
 class JSONFormField(forms.fields.CharField):