+++ feincms/utils/__init__.py
@@ -5,13 +5,13 @@
 from __future__ import absolute_import, division, unicode_literals
 
 import re
+import six
 
 from importlib import import_module
 
 from django.apps import apps
 from django.core.exceptions import ImproperlyConfigured
 from django.db.models import AutoField
-from django.utils import six
 
 from feincms import settings